fix: some other controllers
