+++ Challenge04/Challenge04/Controller/AllTasksController.swift
@@ -17,6 +17,7 @@ class AllTasksController: UIViewController, UITableViewDelegate, UITableViewData
     var goal: CategoryTypes = .none
     
     private var taskModel = [TaskModel]()
+    
     let context = (UIApplication.shared.delegate as! AppDelegate).persistentContainer.viewContext
     
     override func viewDidLoad() {
@@ -50,26 +51,6 @@ class AllTasksController: UIViewController, UITableViewDelegate, UITableViewData
             //error
         }
         
-    }
-    
-    func deleteItem(item: TaskModel){
-        context.delete(item)
-        
-        do {
-            try context.save()
-        } catch {
-            // error
-        }
-    }
-    
-    func updateItem(item: TaskModel, newTitle: String){
-        item.title = newTitle
-        
-        do {
-            try context.save()
-        } catch {
-            // error
-        }
     }
     
     @IBAction func addTasks(_ sender: UIBarButtonItem) {
